pyproject.toml
[project]
name = "wkp"
version = "0.1.0"
description = "CLI tools for Wikipedia download, translation, preview, and publish"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mwclient>=0.11.0",
    "mwparserfromhell>=0.7.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.21.1",
]

[project.scripts]
wkp = "wkp.cli:app"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
